# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.APIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      prefix:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateAPIKeyRequest:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
      permissions:
        items:
          type: string
        type: array
      type:
        enum:
        - private_key
        - publishable_key
        type: string
    required:
    - name
    - type
    type: object
  dto.CreateAPIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        description: Raw key, only returned once
        type: string
      last_used_at:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      prefix:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateReviewRequest:
    properties:
      content:
        maxLength: 3000
        type: string
      entity_id:
        maxLength: 100
        minLength: 1
        type: string
      entity_type:
        allOf:
        - $ref: '#/definitions/types.ReviewEntityType'
        maxLength: 50
      images:
        items:
          type: string
        type: array
      rating:
        maximum: 5
        minimum: 1
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        type: string
    required:
    - entity_id
    - entity_type
    - rating
    type: object
...
  dto.ListAPIKeysResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.APIKeyResponse'
        type: array
      total:
        type: integer
    type: object
...
  dto.RatingStatsResponse:
    properties:
      average_rating:
        type: number
      entity_id:
        type: string
      entity_type:
        $ref: '#/definitions/types.ReviewEntityType'
      five_star_count:
        type: integer
      four_star_count:
        type: integer
      one_star_count:
        type: integer
      rating_distribution:
        additionalProperties:
          type: integer
        description: rating -> count
        type: object
      reviews_with_content:
        type: integer
      reviews_with_images:
        type: integer
      three_star_count:
        type: integer
      total_reviews:
        type: integer
      two_star_count:
        type: integer
      verified_reviews:
        type: integer
    type: object
  dto.ReviewResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      entity_id:
        type: string
      entity_type:
        $ref: '#/definitions/types.ReviewEntityType'
      helpful_count:
        type: integer
      id:
        type: string
      images:
        items:
          type: string
        type: array
      is_featured:
        type: boolean
      is_verified:
        type: boolean
      not_helpful_count:
        type: integer
      rating:
        type: number
      status:
        $ref: '#/definitions/types.Status'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateAPIKeyRequest:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
      permissions:
        items:
          type: string
        type: array
      status:
        allOf:
        - $ref: '#/definitions/types.Status'
        enum:
        - published
        - archived
    type: object
...
  dto.UpdateReviewRequest:
    properties:
      content:
        maxLength: 3000
        type: string
      images:
        items:
          type: string
        type: array
      rating:
        maximum: 5
        minimum: 1
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        type: string
    type: object
...
  ierr.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      internal_error:
        type: string
      message:
        type: string
    type: object
  ierr.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/ierr.ErrorDetail'
      success:
        type: boolean
    type: object
...
  types.ListResponse-dto_ReviewResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ReviewResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.ReviewEntityType:
    enum:
    - place
    - hotel
    - restaurant
    - event
    - experience
    - attraction
    type: string
    x-enum-varnames:
    - EntityTypePlace
    - EntityTypeHotel
    - EntityTypeRestaurant
    - EntityTypeEvent
    - EntityTypeExperience
    - EntityTypeAttraction
  types.Status:
    enum:
    - published
    - draft
    - archived
    - deleted
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDraft
    - StatusArchived
    - StatusDeleted
...
host: 5p9ubi66hh.execute-api.ap-south-1.amazonaws.com
info:
  contact:
    email: support@nashikdarshan.com
    name: API Support
  description: API for Nashik Darshan
  termsOfService: http://nashikdarshan.com/terms/
  title: Nashik Darshan API
  version: "1.0"
paths:
...
  /reviews:
    get:
      consumes:
      - application/json
      description: Get a paginated list of reviews with filtering
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Entity type (place, experience, etc.)
        in: query
        name: entity_type
        type: string
      - description: Entity ID
        in: query
        name: entity_id
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Minimum rating
        in: query
        name: min_rating
        type: number
      - description: Maximum rating
        in: query
        name: max_rating
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListResponse-dto_ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: List reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Create a new review for a place or other entity
      parameters:
      - description: Create review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Create a new review
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a review
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Delete a review
      tags:
      - Reviews
    get:
      consumes:
      - application/json
      description: Get a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get review by ID
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Update an existing review
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Update review request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Update a review
      tags:
      - Reviews
  /reviews/stats/{entityType}/{entityId}:
    get:
      consumes:
      - application/json
      description: Get rating statistics for an entity (place, experience, etc.)
      parameters:
      - description: Entity type (place, experience, etc.)
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RatingStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get rating statistics
      tags:
      - Reviews
  /secrets/api-keys:
    get:
      consumes:
      - application/json
      description: List all API keys for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListAPIKeysResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: List API keys
      tags:
      - Secrets
    post:
      consumes:
      - application/json
      description: Create a new API key (private or publishable). The raw key is returned
        only once.
      parameters:
      - description: Create API key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Create a new API key
      tags:
      - Secrets
  /secrets/api-keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete (archive) an API key
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Delete API key
      tags:
      - Secrets
    get:
      consumes:
      - application/json
      description: Get a specific API key by its ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIKeyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Get API key by ID
      tags:
      - Secrets
    patch:
      consumes:
      - application/json
      description: Update an API key (name, permissions, or status)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      - description: Update API key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Update API key
      tags:
      - Secrets
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key for server-to-server or script access. Use X-API-Key header
      (alternative to Bearer).
    in: header
    name: X-API-Key
    type: apiKey
  Authorization:
    description: JWT Bearer token, e.g. `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"